Extract repeated mock definitions
